pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vl-rag-graph-rlm"
version = "0.1.0"
description = "Vision-Language RAG Graph Recursive Language Models - Multimodal RAG with Qwen3-VL, cheap SOTA LLMs, and graph-based reasoning"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["llm", "ai", "recursive", "rag", "multimodal", "vision-language", "graph", "qwen3vl", "openrouter", "zenmux", "z.ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0",
    "litellm>=1.0.0",
    "RestrictedPython>=7.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "fuzzywuzzy>=0.18.0",
    "numpy>=1.24.0",
    "pymilvus>=2.4.0",
    "jieba>=0.42.0",
    "tiktoken>=0.7.0",
    "networkx>=3.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
qwen3vl = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "pillow>=10.0.0",
    "qwen-vl-utils>=0.0.8",
    "pdf2image>=1.16.0",
    "pymupdf>=1.23.0",
    "python-pptx>=0.6.21",
    "accelerate>=0.27.0",
    "einops>=0.8.0",
]
paddle = [
    "paddlenlp>=2.6.0",
    "pymilvus>=2.4.0",
    "jieba>=0.42.0",
]
all = [
    "vl-rag-graph-rlm[qwen3vl,paddle]",
]

[project.urls]
Homepage = "https://github.com/yourusername/vl-rag-graph-rlm"
Repository = "https://github.com/yourusername/vl-rag-graph-rlm"

[project.scripts]
vrlmrag = "vrlmrag:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
